scanner/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
python-dotenv==1.0.1
scikit-learn==1.5.2
lightgbm==4.5.0
pyarrow==18.1.0

